+++ webpack.config.js
@@ -1,17 +1,15 @@
 const argv = require('yargs').argv;
 
 const isDevelopment = argv.mode === 'development';
-const isTest = argv.mode === 'test';
-const isProduction = !isDevelopment && !isTest;
+const isProduction = !isDevelopment;
 const opts = {
 	isDevelopment,
-	isTest,
 	isProduction
 };
 
 const webpackDevConfig = require('./webpack.dev.config.js')(opts);
 const webpackProdConfig = require('./webpack.prod.config.js')(opts);
-const webpackTestConfig = require('./webpack.test.config.js')(opts);
+const webpackTestConfig = require('./webpack.test.config.js');
 
 const config = Object.assign(isDevelopment
 	? webpackDevConfig